Year 2 (2022-2023)/COMP0010 - Software Engineering/python-shell-p31/requirements.txt
antlr4-python3-runtime
nose2
coverage
flake8==3.8.0
flake8-html
junit2html